[refactor] react.svg 파일관리
