--- src/Home.tsx
@@ -1,3 +1,0 @@
-export default function Home() {
-  return <div>Home</div>;
-}
